requirements.txt
Django>=4.0
djangorestframework
django-environ
psycopg2-binary
Faker
drf-yasg
django-filter
gunicorn
